Implement crude space width determinator
First of all, compute an approximate space width during decomposition of data into text runs, and second, apply some additional magic number logic during post-processing to improve detection of text runs which should be merged.

Presently, this arrangement was arrived at through a binary approach to find numbers and approaches that work as best as possible to get my work done with the hopes I'll return and further tune and test this against inputs outside of my current, and narrow, scope of work.